move buildAttributes to instance root
